configure.ac: make cgroup2 configurable
We would like cgroup2 to be configurable at build time, when it's
disabled, finit will not check /sys/fs/cgroup for cgroup2 support.

The reason for why we need this is because we'd like to use our own
rules for controlling cgroup2, and it's hard to adapt to finit's
default settings.

Signed-off-by: Ming Liu <liu.ming50@gmail.com>